added useAvatar custom hook
